refactor(cli): colocate parsed types with deserialization
Move ParsedWall, ParsedCpu, ParsedAlloc, StableIdentity from types.rs
to report/load.rs where NDJSON deserialization produces them. Inner
fields are now module-private — only the deserializer can construct
parsed values from raw integers.

CorrectedWall, CorrectedCpu, and the bias correction functions stay
in types.rs (the producing module for corrected values). Bias functions
use raw() accessors instead of direct field access.

Test constructors (new_test) gated behind #[cfg(test)].
